+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Types of changes:
 
 ### Added
 
+* Script to convert any input corpus to jsonl files formatted for use by prodigy
+* Add encoding argument for reading corpus (defaults to `latin1` as before)
 * Convert jsonl file to `datasets.DatasetDict` for processing with `transformers` library
 * Function to train a model on the `datasets.DatasetDict` using a pretrained model
 * Log progress when running `get_text_snippets` every 1000 records